+++ airflow/providers/common/sql/operators/sql.py
@@ -25,7 +25,7 @@ from airflow.compat.functools import cached_property
 from airflow.exceptions import AirflowException, AirflowFailException
 from airflow.hooks.base import BaseHook
 from airflow.models import BaseOperator, SkipMixin
-from airflow.providers.common.sql.hooks.sql import DbApiHook, fetch_all_handler
+from airflow.providers.common.sql.hooks.sql import DbApiHook, fetch_all_handler, return_single_query_results
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
@@ -49,6 +49,35 @@ def _parse_boolean(val: str) -> str | bool:
     if val in ("n", "no", "f", "false", "off", "0"):
         return False
     raise ValueError(f"{val!r} is not a boolean-like string value")
+
+
+def _get_failed_checks(checks, col=None):
+    """
+    IMPORTANT!!! Keep it for compatibility with released 8.4.0 version of google provider.
+
+    Unfortunately the provider used _get_failed_checks and parse_boolean as imports and we should
+    keep those methods to avoid 8.4.0 version from failing.
+    """
+    if col:
+        return [
+            f"Column: {col}\nCheck: {check},\nCheck Values: {check_values}\n"
+            for check, check_values in checks.items()
+            if not check_values["success"]
+        ]
+    return [
+        f"\tCheck: {check},\n\tCheck Values: {check_values}\n"
+        for check, check_values in checks.items()
+        if not check_values["success"]
+    ]
+
+
+parse_boolean = _parse_boolean
+"""
+IMPORTANT!!! Keep it for compatibility with released 8.4.0 version of google provider.
+
+Unfortunately the provider used _get_failed_checks and parse_boolean as imports and we should
+keep those methods to avoid 8.4.0 version from failing.
+"""
 
 
 _PROVIDERS_MATCHER = re.compile(r"airflow\.providers\.(.*)\.hooks.*")
@@ -160,11 +189,17 @@ class SQLExecuteQueryOperator(BaseSQLOperator):
     Executes SQL code in a specific database
     :param sql: the SQL code or string pointing to a template file to be executed (templated).
     File must have a '.sql' extensions.
+
+    When implementing a specific Operator, you can also implement `_process_output` method in the
+    hook to perform additional processing of values returned by the DB Hook of yours. For example, you
+    can join description retrieved from the cursors of your statements with returned values, or save
+    the output of your operator to a file.
+
     :param autocommit: (optional) if True, each command is automatically committed (default: False).
     :param parameters: (optional) the parameters to render the SQL query with.
     :param handler: (optional) the function that will be applied to the cursor (default: fetch_all_handler).
     :param split_statements: (optional) if split single SQL string into statements (default: False).
-    :param return_last: (optional) if return the result of only last statement (default: True).
+    :param return_last: (optional) return the result of only last statement (default: True).
 
     .. seealso::
         For more information on how to use this operator, take a look at the guide:
@@ -195,31 +230,42 @@ class SQLExecuteQueryOperator(BaseSQLOperator):
         self.split_statements = split_statements
         self.return_last = return_last
 
+    def _process_output(self, results: list[Any], descriptions: list[Sequence[Sequence] | None]) -> list[Any]:
+        """
+        Processes output before it is returned by the operator.
+
+        It can be overridden by the subclass in case some extra processing is needed. Note that unlike
+        DBApiHook return values returned - the results passed and returned by ``_process_output`` should
+        always be lists of results - each element of the list is a result from a single SQL statement
+        (typically this will be list of Rows). You have to make sure that this is the same for returned
+        values = there should be one element in the list for each statement executed by the hook..
+
+        The "process_output" method can override the returned output - augmenting or processing the
+        output as needed - the output returned will be returned as execute return value and if
+        do_xcom_push is set to True, it will be set as XCom returned.
+
+        :param results: results in the form of list of rows.
+        :param descriptions: list of descriptions returned by ``cur.description`` in the Python DBAPI
+        """
+        return results
+
     def execute(self, context):
         self.log.info("Executing: %s", self.sql)
         hook = self.get_db_hook()
-        if self.do_xcom_push:
-            output = hook.run(
-                sql=self.sql,
-                autocommit=self.autocommit,
-                parameters=self.parameters,
-                handler=self.handler,
-                split_statements=self.split_statements,
-                return_last=self.return_last,
-            )
-        else:
-            output = hook.run(
-                sql=self.sql,
-                autocommit=self.autocommit,
-                parameters=self.parameters,
-                split_statements=self.split_statements,
-            )
-
-        if hasattr(self, "_process_output"):
-            for out in output:
-                self._process_output(*out)
-
-        return output
+        output = hook.run(
+            sql=self.sql,
+            autocommit=self.autocommit,
+            parameters=self.parameters,
+            handler=self.handler if self.do_xcom_push else None,
+            split_statements=self.split_statements,
+            return_last=self.return_last,
+        )
+        if return_single_query_results(self.sql, self.return_last, self.split_statements):
+            # For simplicity, we pass always list as input to _process_output, regardless if
+            # single query results are going to be returned, and we return the first element
+            # of the list in this case from the (always) list returned by _process_output
+            return self._process_output([output], hook.descriptions)[-1]
+        return self._process_output(output, hook.descriptions)
 
     def prepare_template(self) -> None:
         """Parse template file for attribute parameters."""
@@ -231,6 +277,7 @@ class SQLColumnCheckOperator(BaseSQLOperator):
     """
     Performs one or more of the templated checks in the column_checks dictionary.
     Checks are performed on a per-column basis specified by the column_mapping.
+
     Each check can take one or more of the following options:
     - equal_to: an exact value to equal, cannot be used with other comparison options
     - greater_than: value that result should be strictly greater than