+++ airflow/www/static/js/dag/details/taskInstance/index.tsx
@@ -111,7 +111,7 @@ const TaskInstance = ({
 
   const { executionDate } = run;
 
-  let taskActionsTitle = 'Task Actions';
+  let taskActionsTitle = `${isGroup ? 'Task Group' : 'Task'} Actions`;
   if (isMapped) {
     taskActionsTitle += ` for ${actionsMapIndexes.length || 'all'} mapped task${actionsMapIndexes.length !== 1 ? 's' : ''}`;
   }
@@ -174,22 +174,21 @@ const TaskInstance = ({
                   runId={runId}
                   taskId={taskId}
                   mapIndex={instance.mapIndex}
-                  initialValue={instance.notes}
+                  initialValue={instance.note}
                   key={dagId + runId + taskId + instance.mapIndex}
                 />
               )}
-              {!isGroup && (
-                <Box mb={8}>
-                  <TaskActions
-                    title={taskActionsTitle}
-                    runId={runId}
-                    taskId={taskId}
-                    dagId={dagId}
-                    executionDate={executionDate}
-                    mapIndexes={actionsMapIndexes}
-                  />
-                </Box>
-              )}
+              <Box mb={8}>
+                <TaskActions
+                  title={taskActionsTitle}
+                  runId={runId}
+                  taskId={taskId}
+                  dagId={dagId}
+                  executionDate={executionDate}
+                  mapIndexes={actionsMapIndexes}
+                  isGroup={isGroup}
+                />
+              </Box>
               <Details instance={instance} group={group} dagId={dagId} />
               {!isMapped && (
                 <ExtraLinks