+++ tests/providers/databricks/hooks/test_databricks_sql.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,10 +14,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
+#
 from __future__ import annotations
 
-import unittest
 from unittest import mock
+from unittest.mock import patch
 
 import pytest
 
@@ -34,25 +35,156 @@ HOST_WITH_SCHEME = "https://xx.cloud.databricks.com"
 TOKEN = "token"
 
 
-class TestDatabricksSqlHookQueryByName(unittest.TestCase):
-    """
-    Tests for DatabricksHook.
-    """
+@provide_session
+@pytest.fixture(autouse=True)
+def create_connection(session):
+    conn = session.query(Connection).filter(Connection.conn_id == DEFAULT_CONN_ID).first()
+    conn.host = HOST
+    conn.login = None
+    conn.password = TOKEN
+    conn.extra = None
+    session.commit()
 
-    @provide_session
-    def setUp(self, session=None):
-        conn = session.query(Connection).filter(Connection.conn_id == DEFAULT_CONN_ID).first()
-        conn.host = HOST
-        conn.login = None
-        conn.password = TOKEN
-        conn.extra = None
-        session.commit()
 
-        self.hook = DatabricksSqlHook(sql_endpoint_name="Test")
+@pytest.fixture
+def databricks_hook():
+    return DatabricksSqlHook(sql_endpoint_name="Test")
 
-    @mock.patch("airflow.providers.databricks.hooks.databricks_sql.DatabricksSqlHook.get_conn")
-    @mock.patch("airflow.providers.databricks.hooks.databricks_base.requests")
-    def test_query(self, mock_requests, mock_conn):
+
+def get_cursor_descriptions(fields: list[str]) -> list[tuple[str]]:
+    return [(field,) for field in fields]
+
+
+@pytest.mark.parametrize(
+    "return_last, split_statements, sql, cursor_calls,"
+    "cursor_descriptions, cursor_results, hook_descriptions, hook_results, ",
+    [
+        pytest.param(
+            True,
+            False,
+            "select * from test.test",
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[1, 2], [11, 12]],
+            id="The return_last set and no split statements set on single query in string",
+        ),
+        pytest.param(
+            False,
+            False,
+            "select * from test.test;",
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[1, 2], [11, 12]],
+            id="The return_last not set and no split statements set on single query in string",
+        ),
+        pytest.param(
+            True,
+            True,
+            "select * from test.test;",
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[1, 2], [11, 12]],
+            id="The return_last set and split statements set on single query in string",
+        ),
+        pytest.param(
+            False,
+            True,
+            "select * from test.test;",
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[[1, 2], [11, 12]]],
+            id="The return_last not set and split statements set on single query in string",
+        ),
+        pytest.param(
+            True,
+            True,
+            "select * from test.test;select * from test.test2;",
+            ["select * from test.test", "select * from test.test2"],
+            [["id", "value"], ["id2", "value2"]],
+            ([[1, 2], [11, 12]], [[3, 4], [13, 14]]),
+            [[("id2",), ("value2",)]],
+            [[3, 4], [13, 14]],
+            id="The return_last set and split statements set on multiple queries in string",
+        ),
+        pytest.param(
+            False,
+            True,
+            "select * from test.test;select * from test.test2;",
+            ["select * from test.test", "select * from test.test2"],
+            [["id", "value"], ["id2", "value2"]],
+            ([[1, 2], [11, 12]], [[3, 4], [13, 14]]),
+            [[("id",), ("value",)], [("id2",), ("value2",)]],
+            [[[1, 2], [11, 12]], [[3, 4], [13, 14]]],
+            id="The return_last not set and split statements set on multiple queries in string",
+        ),
+        pytest.param(
+            True,
+            True,
+            ["select * from test.test;"],
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[[1, 2], [11, 12]]],
+            id="The return_last set on single query in list",
+        ),
+        pytest.param(
+            False,
+            True,
+            ["select * from test.test;"],
+            ["select * from test.test"],
+            [["id", "value"]],
+            ([[1, 2], [11, 12]],),
+            [[("id",), ("value",)]],
+            [[[1, 2], [11, 12]]],
+            id="The return_last not set on single query in list",
+        ),
+        pytest.param(
+            True,
+            True,
+            "select * from test.test;select * from test.test2;",
+            ["select * from test.test", "select * from test.test2"],
+            [["id", "value"], ["id2", "value2"]],
+            ([[1, 2], [11, 12]], [[3, 4], [13, 14]]),
+            [[("id2",), ("value2",)]],
+            [[3, 4], [13, 14]],
+            id="The return_last set set on multiple queries in list",
+        ),
+        pytest.param(
+            False,
+            True,
+            "select * from test.test;select * from test.test2;",
+            ["select * from test.test", "select * from test.test2"],
+            [["id", "value"], ["id2", "value2"]],
+            ([[1, 2], [11, 12]], [[3, 4], [13, 14]]),
+            [[("id",), ("value",)], [("id2",), ("value2",)]],
+            [[[1, 2], [11, 12]], [[3, 4], [13, 14]]],
+            id="The return_last not set on multiple queries not set",
+        ),
+    ],
+)
+def test_query(
+    databricks_hook,
+    return_last,
+    split_statements,
+    sql,
+    cursor_calls,
+    cursor_descriptions,
+    cursor_results,
+    hook_descriptions,
+    hook_results,
+):
+    with patch(
+        "airflow.providers.databricks.hooks.databricks_sql.DatabricksSqlHook.get_conn"
+    ) as mock_conn, patch("airflow.providers.databricks.hooks.databricks_base.requests") as mock_requests:
         mock_requests.codes.ok = 200
         mock_requests.get.return_value.json.return_value = {
             "endpoints": [
@@ -68,26 +200,41 @@ class TestDatabricksSqlHookQueryByName(unittest.TestCase):
         }
         status_code_mock = mock.PropertyMock(return_value=200)
         type(mock_requests.get.return_value).status_code = status_code_mock
+        connections = []
+        cursors = []
+        for index in range(len(cursor_descriptions)):
+            conn = mock.MagicMock()
+            cur = mock.MagicMock(
+                rowcount=len(cursor_results[index]),
+                description=get_cursor_descriptions(cursor_descriptions[index]),
+            )
+            cur.fetchall.return_value = cursor_results[index]
+            conn.cursor.return_value = cur
+            cursors.append(cur)
+            connections.append(conn)
+        mock_conn.side_effect = connections
+        results = databricks_hook.run(
+            sql=sql, handler=fetch_all_handler, return_last=return_last, split_statements=split_statements
+        )
 
-        test_fields = ["id", "value"]
-        test_schema = [(field,) for field in test_fields]
+        assert databricks_hook.descriptions == hook_descriptions
+        assert databricks_hook.last_description == hook_descriptions[-1]
+        assert results == hook_results
 
-        conn = mock_conn.return_value
-        cur = mock.MagicMock(rowcount=0, description=test_schema)
-        cur.fetchall.return_value = []
-        conn.cursor.return_value = cur
-
-        query = "select * from test.test;"
-        schema, results = self.hook.run(sql=query, handler=fetch_all_handler)
-
-        assert schema == test_schema
-        assert results == []
-
-        cur.execute.assert_has_calls([mock.call(q) for q in [query]])
+        for index, cur in enumerate(cursors):
+            cur.execute.assert_has_calls([mock.call(cursor_calls[index])])
         cur.close.assert_called()
 
-    def test_no_query(self):
-        for empty_statement in ([], "", "\n"):
-            with pytest.raises(ValueError) as err:
-                self.hook.run(sql=empty_statement)
-            assert err.value.args[0] == "List of SQL statements is empty"
+
+@pytest.mark.parametrize(
+    "empty_statement",
+    [
+        pytest.param([], id="Empty list"),
+        pytest.param("", id="Empty string"),
+        pytest.param("\n", id="Only EOL"),
+    ],
+)
+def test_no_query(databricks_hook, empty_statement):
+    with pytest.raises(ValueError) as err:
+        databricks_hook.run(sql=empty_statement)
+    assert err.value.args[0] == "List of SQL statements is empty"